Fixed UNIX in login to be f64 to comply with javascript Date.now()

+++ server/src/api/v1/handlers/login.rs
@@ -15,7 +15,7 @@ use serde::{Deserialize, Serialize};
 pub struct LoginForm<'r> {
     username: &'r str,
     pw: &'r str,
-    iat: u32,
+    iat: f64,
 }
 
 /**
